+++ BulkCommand.php
@@ -72,15 +72,15 @@ class BulkCommand extends Component
             $body = '{}';
         } elseif (is_array($this->actions)) {
             $body = '';
-            $prettyPrintSupport = property_exists('yii\\helpers\\Json', 'prettyPrint');
-            if ($prettyPrintSupport) {
+            $prettyPrintSupported = property_exists('yii\\helpers\\Json', 'prettyPrint');
+            if ($prettyPrintSupported) {
                 $originalPrettyPrint = Json::$prettyPrint;
                 Json::$prettyPrint = false; // ElasticSearch bulk API uses new lines as delimiters.
             }
             foreach ($this->actions as $action) {
                 $body .= Json::encode($action) . "\n";
             }
-            if ($prettyPrintSupport) {
+            if ($prettyPrintSupported) {
                 Json::$prettyPrint = $originalPrettyPrint;
             }
         } else {