+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 Elasticsearch extension Change Log
 2.1.5 under development
 -----------------------
 
-- no changes in this release.
+- Bug #344: Disabled JSON pretty print for ElasticSearch bulk API (rhertogh)
 
 
 2.1.4 May 22, 2023